Add files via upload
The Iris Dataset can be downloaded from the UCI ML Repository—
Download Iris Flowers Dataset. The goal of this data science project for
beginners is to classify the flowers into three species—Virginia, setosa, or
versicolor—based on the length and width of the petals and sepals. By
implementing advanced algorithms, you can also add this project to your deep
learning projects portfolio.